refactor TableMove to separate file
